requirements.txt
# Core Azure Functions and services
azure-functions
azure-storage-blob>=12.0.0
azure-ai-textanalytics>=5.3.0

# Testing dependencies
requests
python-dotenv

# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry
